pyproject.toml
[project]
name = "invoice-agent"
version = "0.1.0"
description = "Automated invoice extraction agent using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.13.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
]
